+++ openquake/commands/show.py
@@ -22,7 +22,8 @@ import os
 import logging
 
 from openquake.hazardlib.calc.hazard_curve import zero_curves
-from openquake.commonlib import sap, datastore
+from openquake.baselib import sap
+from openquake.commonlib import datastore
 from openquake.commonlib.writers import write_csv
 from openquake.commonlib.util import rmsep
 from openquake.risklib import scientific
