+++ openquake/commands/tidy.py
@@ -17,7 +17,8 @@
 # along with OpenQuake. If not, see <http://www.gnu.org/licenses/>.
 
 from __future__ import print_function
-from openquake.commonlib import sap, nrml, writers
+from openquake.baselib import sap
+from openquake.commonlib import nrml, writers
 
 
 @sap.Script
